requirements.txt
streamlit==1.30.0
langchain==0.1.2
openai==1.9.0
numpy==1.26.3
pathlib==1.0.1
ffmpeg==1.4
streamlit-audiorecorder==0.0.4
langchain-openai==0.0.3